Building a framework
