+++ src/main.rs
@@ -2,24 +2,24 @@
 #![no_main]
 #![windows_subsystem = "windows"]
 
-// pub mod audio;
+pub mod audio;
 pub mod visual;
 
-use intro_rs::audio::Audio;
-use intro_rs::intro::Intro;
-use intro_rs::visual::Visual;
+use intro_rs::Audio;
+use intro_rs::Intro;
+use intro_rs::Visual;
 
 pub struct Namekusei {
     time: f32,
-    audio: Audio,
-    visual: Visual
+    audio: audio::Audio,
+    visual: visual::Visual
 }
 
 impl Intro for Namekusei {
     fn new() -> Self where Self: Sized {
         let time = 0.0;
-        let audio = Audio::new();
-        let visual = Visual::new();
+        let audio = audio::Audio::new();
+        let visual = visual::Visual::new();
         Self { time, audio, visual }
     }
 
@@ -27,11 +27,11 @@ impl Intro for Namekusei {
         &mut self.time
     }
 
-    fn audio(&self) -> &Audio {
+    fn audio(&self) -> &dyn Audio {
         &self.audio
     }
 
-    fn visual(&self) -> &Visual {
+    fn visual(&self) -> &dyn Visual {
         &self.visual
     }
 }
